JsApiService: Reorganize
